Fix date parsing logic
There was a problem where if a date that was just a year such as
"2017" was passed into the date parsing logic it would be interpreted
as ms since the epoch instead of going through the date parser.
This has been fixed along with some subtle cleanups to names of
methods and doc strings to make the code more readable.